Create linear export for GNPS fbmn remove concurrency.

+++ src/main/java/io/github/mzmine/modules/io/export_features_gnps/fbmn/GnpsFbmnMgfExportTask.java
@@ -57,7 +57,7 @@ import java.util.Arrays;
 import java.util.ConcurrentModificationException;
 import java.util.List;
 import java.util.Objects;
-import java.util.concurrent.atomic.AtomicLong;
+import java.util.concurrent.atomic.AtomicInteger;
 import java.util.logging.Level;
 import java.util.logging.Logger;
 import java.util.regex.Pattern;
@@ -82,7 +82,7 @@ public class GnpsFbmnMgfExportTask extends AbstractTask implements ProcessedItem
   private final boolean mergeMS2;
   private final FeatureListRowsFilter filter;
   // track number of exported items
-  private final AtomicLong exportedRows = new AtomicLong(0);
+  private final AtomicInteger exportedRows = new AtomicInteger(0);
   private int currentIndex = 0;
   // by robin
   private NumberFormat mzForm = MZmineCore.getConfiguration().getMZFormat();
@@ -312,7 +312,7 @@ public class GnpsFbmnMgfExportTask extends AbstractTask implements ProcessedItem
   }
 
   @Override
-  public long getProcessedItems() {
+  public int getProcessedItems() {
     return exportedRows.get();
   }
 }
